backend/requirements.txt
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
async-timeout==5.0.1
attrs==25.4.0
Brotli==1.1.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
clerk-sdk-python==0.1.0
clerk-django==1.0.3
Deprecated==1.2.18
Django==4.2.25
django-cors-headers==4.9.0
djangorestframework==3.16.1
exceptiongroup==1.3.0
filelock==3.19.1
frozenlist==1.8.0
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.3
mpmath==1.3.0
multidict==6.7.0
networkx==3.2.1
numpy==2.0.2
packaging==25.0
pillow==11.3.0
propcache==0.4.1
psycopg2-binary==2.9.11
PyJWT==2.8.0
pycares==4.11.0
pycparser==2.23
pydantic==2.12.0
pydantic_core==2.41.1
pyngrok==7.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==5.1.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
svix==1.77.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.8.0
tqdm==4.67.1
transformers==4.57.0
types-Deprecated==1.2.15.20250304
types-python-dateutil==2.9.0.20251008
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
wrapt==1.17.3
yarl==1.22.0
zstandard==0.25.0
